% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.r
\name{add_filter}
\alias{add_filter}
\title{Add a filter}
\usage{
add_filter(vl, expr)
}
\arguments{
\item{vl}{Vega-Lite object created by \code{\link{vegalite}}}

\item{expr}{Vega Expression for filtering data items (or rows). Each datum
object can be referred using bound variable datum. For example, setting
\code{expr} to \code{"datum.datum.b2 > 60"} would make the output data includes only
items that have values in the field \code{b2} over 60.}
}
\description{
Add a filter
}

