% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.r
\encoding{UTF-8}
\name{axis_facet_row}
\alias{axis_facet_row}
\title{General axis setttings (row facets)}
\usage{
axis_facet_row(vl, ...)
}
\arguments{
\item{vl}{Vega-Lite object}

\item{axisColor, axisWidth, offset, orient, zindex}{see \href{https://vega.github.io/vega-lite/docs/axis.html#general}{axis docs} &
\href{https://github.com/vega/vega-lite/blob/master/src/axis.ts}{axis base config}}

\item{grid, gridColor, gridDash, gridOpacity, gridWidth}{see \href{https://vega.github.io/vega-lite/docs/axis.html#grid}{axis docs} &
\href{https://github.com/vega/vega-lite/blob/master/src/axis.ts}{axis base config}}

\item{labels, format, labelAngle, labelMaxLength, shortTimeLabels}{see \href{https://vega.github.io/vega-lite/docs/axis.html#labels}{axis docs} &
\href{https://github.com/vega/vega-lite/blob/master/src/axis.ts}{axis base config}}

\item{subdivide, ticks, tickColor, tickLabelColor, tickLabelFont, tickLabelFontSize, tickPadding, tickSize, tickSizeMajor, tickSizeMinor, tickSizeEnd, tickWidth, values}{see \href{https://vega.github.io/vega-lite/docs/axis.html#ticks}{axis docs} &
\href{https://github.com/vega/vega-lite/blob/master/src/axis.ts}{axis base config}}

\item{title, titleColor, titleFont, titleFontWeight, titleFontSize, titleOffset, titleMaxLength, characterWidth}{see \href{https://vega.github.io/vega-lite/docs/axis.html#title}{axis docs} &
\href{https://github.com/vega/vega-lite/blob/master/src/axis.ts}{axis base config}}

\item{remove}{see \href{http://vega.github.io/vega-lite/docs/axis.html}{axis docs} &
\href{https://github.com/vega/vega-lite/blob/master/src/axis.ts}{axis base config}}
}
\description{
Axes provide axis lines, ticks and labels to convey how a spatial range represents
a data range. Simply put, axes visualize scales. \cr
\cr
By default, Vega-Lite automatically creates axes for x, y, row, and column channels
when they are encoded. Axis can be customized via the axis property of a channel
definition.
}
\references{
\href{http://vega.github.io/vega-lite/docs/axis.html}{Vega-List Axis spec}
}

