% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.r
\name{axis_grid_y}
\alias{axis_grid_y}
\title{Axis grid settings}
\usage{
axis_grid_y(vl, has_grid)
}
\description{
Axis grid settings
}
\references{
\href{http://vega.github.io/vega-lite/docs/axis.html}{Vega-List Axis spec}
}

