% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.r
\name{axis_labels_x}
\alias{axis_labels_x}
\title{Axis label settings}
\usage{
axis_labels_x(vl, labels, format, label_angle, label_max_length,
  short_time_labels)
}
\description{
Axis label settings
}
\references{
\href{http://vega.github.io/vega-lite/docs/axis.html}{Vega-List Axis spec}
}

