% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.r
\name{axis_ticks_x}
\alias{axis_ticks_x}
\title{Axis tick settings}
\usage{
axis_ticks_x(vl, values, subdivide, ticks, tick_padding, tick_size,
  tick_size_major, tick_size_minor, tick_size_end)
}
\description{
Axis tick settings
}
\references{
\href{http://vega.github.io/vega-lite/docs/axis.html}{Vega-List Axis spec}
}

