% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.r
\name{axis_title_y}
\alias{axis_title_y}
\title{Axis title settings}
\usage{
axis_title_y(vl, title, title_offset, title_max_length, character_width)
}
\description{
Axis title settings
}
\references{
\href{http://vega.github.io/vega-lite/docs/axis.html}{Vega-List Axis spec}
}

