% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.r
\encoding{UTF-8}
\name{axis_vl}
\alias{axis_vl}
\alias{axis_x}
\alias{axis_y}
\alias{axis_facet_col}
\alias{axis_facet_row}
\title{General axis settings (all)}
\usage{
axis_vl(vl, chnl = "x", domain = NULL, grid = NULL, maxExtent = NULL,
  minExtent = NULL, orient = NULL, offset = NULL, position = NULL,
  zindex = NULL, format = NULL, labels = NULL, labelAngle = NULL,
  labelOverlap = NULL, labelPadding = NULL, ticks = NULL,
  tickCount = NULL, tickSize = NULL, values = NULL, title = NULL,
  titleMaxLength = NULL, titlePadding = NULL, remove = FALSE, ...)

axis_x(vl, ...)

axis_y(vl, ...)

axis_facet_col(vl, ...)

axis_facet_row(vl, ...)
}
\arguments{
\item{vl}{Vega-Lite object}

\item{chnl}{x, y, column, or row}

\item{domain, grid, maxExtent, minExtent, orient, offset, position, zindex}{see \href{https://vega.github.io/vega-lite/docs/axis.html#general}{axis docs}}

\item{format, labels, labelAngle, labelOverlap, labelPadding}{see \href{https://vega.github.io/vega-lite/docs/axis.html#labels}{axis docs}}

\item{ticks, tickCount, tickSize, values}{see \href{https://vega.github.io/vega-lite/docs/axis.html#ticks}{axis docs}}

\item{title, titleMaxLength, titlePadding}{see \href{https://vega.github.io/vega-lite/docs/axis.html#title}{axis docs}}

\item{remove}{see \href{http://vega.github.io/vega-lite/docs/axis.html}{axis docs}}

\item{...}{deprecated arguments}
}
\description{
axis_vl provide axis lines, ticks and labels to convey how a spatial range represents
a data range. Simply put, axes visualize scales. \cr
\cr
By default, Vega-Lite automatically creates axes for x, y, row, and column channels
when they are encoded. Axis can be customized via the axis property of a channel
definition. \cr
\cr
The axis_vl function works with all possible channels, but axis_x, axis_y, axis_facet_row
and axis_facet_col are offered as conveniences.
}
\examples{
vegalite() \%>\%
  add_data("https://vega.github.io/vega-editor/app/data/population.json") \%>\%
  add_filter("datum.year == 2000") \%>\%
  calculate("gender", 'datum.sex == 2 ? "Female" : "Male"') \%>\%
  encode_x("gender", "nominal") \%>\%
  encode_y("people", "quantitative", aggregate="sum") \%>\%
  encode_color("gender", "nominal") \%>\%
  scale_x_ordinal_vl(range_step=8) \%>\%
  scale_color_nominal_vl(range=c("#EA98D2", "#659CCA")) \%>\%
  facet_col("age", "ordinal", padding=4) \%>\%
  axis_x(remove=TRUE) \%>\%
  axis_y(title="population", grid=FALSE) \%>\%
  view_config(stroke_width=0) \%>\%
  mark_bar()

vegalite() \%>\%
  add_data("https://vega.github.io/vega-editor/app/data/population.json") \%>\%
  add_filter("datum.year == 2000") \%>\%
  calculate("gender", 'datum.sex == 2 ? "Female" : "Male"') \%>\%
  encode_x("gender", "nominal") \%>\%
  encode_y("people", "quantitative", aggregate="sum") \%>\%
  encode_color("gender", "nominal") \%>\%
  scale_x_ordinal_vl(range_step=8) \%>\%
  scale_color_nominal_vl(range=c("#EA98D2", "#659CCA")) \%>\%
  facet_col("age", "ordinal", padding=4) \%>\%
  axis_x(remove=TRUE) \%>\%
  axis_y(title="population", grid=FALSE) \%>\%
  view_config(stroke_width=0) \%>\%
  mark_bar()
}
\references{
\href{http://vega.github.io/vega-lite/docs/axis.html}{Vega-List Axis spec}
}
