% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.r
\name{axis_y}
\alias{axis_y}
\title{General axis setttings}
\usage{
axis_y(layer, offset, orient)
}
\description{
General axis setttings
}
\references{
\href{http://vega.github.io/vega-lite/docs/axis.html}{Vega-List Axis spec}
}

