% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.r
\encoding{UTF-8}
\name{config_color}
\alias{config_color}
\title{Color config}
\usage{
config_color(vl, color = NULL, fill = NULL, stroke = NULL)
}
\arguments{
\item{vl}{a Vega-Lite object}

\item{color}{color of the mark – either fill or stroke color based on the filled mark config.}

\item{fill}{fill color. This config will be overridden by color channel’s
specified or mapped values if filled is true.}

\item{stroke}{stroke color. This config will be overridden by color channel’s
specified or mapped values if filled is false.}
}
\description{
Color config
}

