% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.r
\encoding{UTF-8}
\name{config_font}
\alias{config_font}
\title{Font config}
\usage{
config_font(vl, font = NULL, font_size = NULL, font_style = NULL,
  font_weight = NULL)
}
\arguments{
\item{vl}{a Vega-Lite object}

\item{font}{typeface to set the text in (e.g., Helvetica Neue).}

\item{font_size}{font size, in pixels. The default value is 10.}

\item{font_style}{font style (e.g., italic).}

\item{font_weight}{font weight (e.g., bold).}
}
\description{
Font config
}
