% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.r
\encoding{UTF-8}
\name{config_opacity}
\alias{config_opacity}
\title{Opacity config}
\usage{
config_opacity(vl, opacity = NULL, fill_opacity = NULL,
  stroke_opacity = NULL)
}
\arguments{
\item{vl}{a Vega-Lite object}

\item{opacity}{\code{0.0}-\code{1.0}}

\item{fill_opacity}{\code{0.0}-\code{1.0}}

\item{stroke_opacity}{\code{0.0}-\code{1.0}}
}
\description{
Opacity config
}
