% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.r
\name{config_stroke}
\alias{config_stroke}
\title{Stroke config}
\usage{
config_stroke(vl, stroke = NULL, stroke_width = NULL, stroke_dash = NULL,
  stroke_dash_offset = NULL)
}
\arguments{
\item{vl}{a Vega-Lite object}

\item{stroke}{stroke color}

\item{stroke_width}{stroke of the width in pixels}

\item{stroke_dash}{an array of alternating stroke, space lengths for creating
dashed or dotted lines.}

\item{stroke_dash_offset}{the offset (in pixels) into which to begin drawing with the stroke dash array.}

\item{stroke_opacity}{\code{0.0}-\code{1.0}}
}
\description{
Stroke config
}

