% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.r
\encoding{UTF-8}
\name{config_text}
\alias{config_text}
\title{Text config}
\usage{
config_text(vl, angle = NULL, align = NULL, baseline = NULL, dx = NULL,
  dy = NULL, radius = NULL, theta = NULL, format = NULL,
  short_time_labels = NULL, opacity = NULL)
}
\arguments{
\item{vl}{a Vega-Lite object}

\item{angle}{rotation angle of the text, in degrees.}

\item{align}{horizontal alignment of the text. One of left, right, center.}

\item{baseline}{vertical alignment of the text. One of top, middle, bottom.}

\item{dx, dy}{horizontal/vertical in pixels, between the text label and its
anchor point. The offset is applied after rotation by the angle property.}

\item{radius}{polar coordinate radial offset, in pixels, of the text label
from the origin determined by the x and y properties.}

\item{theta}{polar coordinate angle, in radians, of the text label from the
origin determined by the x and y properties. Values for theta follow
the same convention of arc mark startAngle and endAngle properties:
angles are measured in radians, with 0 indicating “north”.}

\item{format}{ormatting pattern for text value. If not defined, this will be
determined automatically}

\item{short_time_labels}{whether month names and weekday names should be abbreviated.}

\item{opacity}{0-1}
}
\description{
Text config
}
\references{
\href{http://vega.github.io/vega-lite/docs/mark.html}{Vega-Lite Mark spec}
}

