% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.r
\encoding{UTF-8}
\name{cell_config}
\alias{cell_config}
\alias{cell_size}
\alias{facet_cell}
\alias{grid_facet}
\title{Deprecated \code{cell} functions}
\usage{
cell_config(vl, ...)

cell_size(vl, width = 200, height = 200, facet = FALSE)

facet_cell(vl, ...)

grid_facet(vl, grid_color = NULL, grid_opacity = NULL, grid_offset = NULL)
}
\arguments{
\item{vl}{Vega-Lite object}

\item{...}{additional arguments to pass to cell_config}

\item{width}{width of cell}

\item{height}{height of cell}

\item{facet}{if facetting}

\item{grid_color}{color of the grid between facets.}

\item{grid_opacity}{\code{0.0}-\code{1.0}}

\item{grid_offset}{offset for grid between facets.}
}
\description{
Vega-lite 2 no longer has \code{cell} and \code{facet} properties under \code{config}.  Thus
functions setting those properties have been deprecated. Some of the
functionality can be acheived with \code{\link{view_config}} and
\code{\link{view_size}}.
}
\references{
\href{http://vega.github.io/vega-lite/docs/config.html}{Vega-Lite 2 config spec}
\href{http://vega.github.io/vega-lite-v1/docs/config.html#cell-config}{Vega-Lite 1 cell spec}
\href{http://vega.github.io/vega-lite-v1/docs/config.html#facet-config}{Vega-Lite 1 cell spec}
}
