% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.r
\name{encode_detail}
\alias{encode_detail}
\title{Encode detail "channel"}
\usage{
encode_detail(vl, field = NULL, type)
}
\arguments{
\item{vl}{Vega-Lite object created by \code{\link{vegalite}}}

\item{field}{single element character vector naming the column}

\item{type}{the encoded field’s type of measurement. This can be either a full type
name (\code{quantitative}, \code{temporal}, \code{ordinal}, and \code{nominal})
or an initial character of the type name (\code{Q}, \code{T}, \code{O}, \code{N}).
This property is case insensitive. If \code{auto} is used, the type will
be guessed (so you may want to actually specify it if you want consistency).}

\item{vl}{Vega-Lite object created by \code{\link{vegalite}}}
}
\description{
Encode detail "channel"
}
\references{
\href{http://vega.github.io/vega-lite/docs/encoding.html}{Vega-Lite Encoding spec}
}

