% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.r
\name{encode_x}
\alias{encode_x}
\title{Encode x "channel"}
\usage{
encode_x(vl, field, type = "auto")
}
\arguments{
\item{vl}{Vega-Lite object created by \code{\link{vegalite}}}

\item{field}{single element character vector naming the column}

\item{type}{the encoded field’s type of measurement. This can be either a full type
name (\code{quantitative}, \code{temporal}, \code{ordinal}, and \code{nominal})
or an initial character of the type name (\code{Q}, \code{T}, \code{O}, \code{N}).
This property is case insensitive. If \code{auto} is used, the type will
be guessed (so you may want to actually specify it if you want consistency).}
}
\description{
Vega-Lite has many "encoding channels". Each channel definition object must
describe the data field encoded by the channel and its data type, or a constant
value directly mapped to the mark properties. In addition, it can describe the
mapped field’s transformation and properties for its scale and guide.
}
\references{
\href{http://vega.github.io/vega-lite/docs/encoding.html}{Vega-Lite Encoding spec}
}

