% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.r
\encoding{UTF-8}
\name{facet_col}
\alias{facet_col}
\title{Create a horizontal ribbon of panels}
\usage{
facet_col(vl, field, type, round = TRUE, padding = 16)
}
\arguments{
\item{vl}{Vega-Lite object}

\item{field}{single element character vector naming the column.}

\item{type}{the encoded field’s type of measurement.}

\item{round}{round values}

\item{padding}{facet padding}
}
\description{
Create a horizontal ribbon of panels
}
\references{
\href{http://vega.github.io/vega-lite/docs/facet.html}{Vega-Lite Faceting}
}

