% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.r
\name{filter_null}
\alias{filter_null}
\title{Filter 'null' values}
\usage{
filter_null(vl, setting = NULL)
}
\arguments{
\item{vl}{Vega-Lite object created by \code{\link{vegalite}}}

\item{setting}{if \code{NULL} only quantitative and temporal fields are
filtered. If \code{TRUE}, all data items with 'null' values are
filtered. If \code{FALSE}, all data items are included.}
}
\description{
Whether to filter null values from the data.
}

