% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.r
\encoding{UTF-8}
\name{grid_facet}
\alias{grid_facet}
\title{Facet grid aesthetics}
\usage{
grid_facet(vl, grid_color = NULL, grid_opacity = NULL, grid_offset = NULL)
}
\arguments{
\item{vl}{Vega-Lite object}

\item{grid_color}{color of the grid between facets.}

\item{grid_opacity}{\code{0.0}-\code{1.0}}

\item{grid_offset}{offset for grid between facets.}
}
\description{
Facet grid aesthetics
}
\references{
\href{http://vega.github.io/vega-lite/docs/config.html#facet-config}{Vega-Lite Facet spec}
}

