% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.r
\encoding{UTF-8}
\name{legend}
\alias{legend}
\title{Legend settings (all)}
\usage{
legend(vl, chnl = "color", orient = NULL, offset = NULL, values = NULL,
  format = NULL, labelAlign = NULL, labelBaseline = NULL,
  labelColor = NULL, labelFont = NULL, labelFontSize = NULL,
  short_time_labels = NULL, symbolColor = NULL, symbolShape = NULL,
  symbolSize = NULL, symbolStrokeWidth = NULL, title = NULL,
  titleColor = NULL, titleFont = NULL, titleFontSize = NULL,
  titleFontWeight = NULL, remove = FALSE)
}
\arguments{
\item{vl}{a Vega-Lite object}

\item{chnl}{a channel, by default vegalite creates legends for color, size, and shape}

\item{orient, offset, values}{see \href{https://vega.github.io/vega-lite/docs/legend.html#general}{legend docs}}

\item{format, labelAlign, labelBaseline, labelColor, labelFont, labelFontSize, short_time_labels}{see \href{https://vega.github.io/vega-lite/docs/legend.html#labels}{legend docs}}

\item{symbolColor, symbolShape, symbolSize, symbolStrokeWidth}{see \href{https://vega.github.io/vega-lite/docs/legend.html#symbols}{legend docs}}

\item{title}{the title for the legend.}

\item{remove}{if \code{TRUE}, there will be no legend for this aesthetic.}

\item{title, titleColor, titleFont, titleFontSize, titleFontWeight}{see \href{https://vega.github.io/vega-lite/docs/legend.html#title}{legend docs}}
}
\description{
Legend settings (all)
}

