% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.r
\encoding{UTF-8}
\name{legend_shape}
\alias{legend_shape}
\title{Legend settings (shape)}
\usage{
legend_shape(vl, orient = NULL, title = NULL, format = NULL,
  short_time_labels = NULL, value = NULL, remove = FALSE)
}
\arguments{
\item{vl}{a Vega-Lite object}

\item{orient}{the orientation of the legend. One of "left" or "right". This
determines how the legend is positioned within the scene.}

\item{title}{the title for the legend.}

\item{format}{the formatting pattern for axis labels. This is D3’s number
format pattern for quantitative axis and D3’s time format pattern
for time axis.}

\item{short_time_labels}{whether month and day names should be abbreviated.}

\item{value}{explicitly set the visible legend values.}

\item{remove}{if \code{TRUE}, there will be no legend for this aesthetic.}
}
\description{
Legend settings (shape)
}

