% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.r
\name{mark_square}
\alias{mark_square}
\title{Square mark}
\usage{
mark_square(vl, size = 30, opacity = NULL)
}
\arguments{
\item{size}{def}

\item{opacity}{0.0-1.0}
}
\description{
Square mark
}
\references{
\href{http://vega.github.io/vega-lite/docs/mark.html}{Vega-Lite Mark spec}
}

