% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.r
\encoding{UTF-8}
\name{mark_text}
\alias{mark_text}
\title{Text mark}
\usage{
mark_text(vl, size = NULL, opacity = NULL)
}
\arguments{
\item{vl}{Vega-Lite object}

\item{size}{The pixel area each the point. For example: in the case of circles,
the radius is determined in part by the square root of the size value.}

\item{opacity}{0.0-1.0}
}
\description{
A text mark represents each data point with a text instead of a point.
}
\references{
\href{http://vega.github.io/vega-lite/docs/mark.html}{Vega-Lite Mark spec}
}

