% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.r
\name{renderVegalite}
\alias{renderVegalite}
\title{Widget render function for use in Shiny}
\usage{
renderVegalite(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

