% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.r
\name{scale_color_nominal}
\alias{scale_color_nominal}
\title{Nominal Color Scale}
\usage{
scale_color_nominal(vl, range)
}
\description{
Nominal Color Scale
}
\references{
\href{http://vega.github.io/vega-lite/docs/scale.html}{Vega-Lite Scales spec}
}

