% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.r
\name{scale_color_sequential}
\alias{scale_color_sequential}
\title{Sequential Color Scale}
\usage{
scale_color_sequential(vl, range)
}
\description{
Sequential Color Scale
}
\references{
\href{http://vega.github.io/vega-lite/docs/scale.html}{Vega-Lite Scales spec}
}

