% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.r
\name{scale_shape}
\alias{scale_shape}
\title{Shape Scale}
\usage{
scale_shape(vl, range)
}
\description{
Shape Scale
}
\references{
\href{http://vega.github.io/vega-lite/docs/scale.html}{Vega-Lite Scales spec}
}

