% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.r
\name{scale_x_quantile}
\alias{scale_x_quantile}
\title{Quantile Scale}
\usage{
scale_x_quantile(vl, domain, range)
}
\description{
Quantile Scale
}
\references{
\href{http://vega.github.io/vega-lite/docs/scale.html}{Vega-Lite Scales spec}
}

