% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.r
\name{scale_x_quantize}
\alias{scale_x_quantize}
\title{Quantize Scale}
\usage{
scale_x_quantize(vl, domain, rang, nice, zeroe)
}
\description{
Quantize Scale
}
\references{
\href{http://vega.github.io/vega-lite/docs/scale.html}{Vega-Lite Scales spec}
}

