% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.r
\name{scale_x_sqrt}
\alias{scale_x_sqrt}
\title{Sqrt Scale}
\usage{
scale_x_sqrt(vl, domain, range, clamp, nice, zero)
}
\description{
Sqrt Scale
}
\references{
\href{http://vega.github.io/vega-lite/docs/scale.html}{Vega-Lite Scales spec}
}

