% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.r
\name{scale_x_time}
\alias{scale_x_time}
\title{Temporal Scale}
\usage{
scale_x_time(vl, domain = NULL, range = NULL, clamp = NULL, nice = NULL)
}
\description{
Temporal Scale
}
\references{
\href{http://vega.github.io/vega-lite/docs/scale.html}{Vega-Lite Scales spec}
}

