% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.r
\name{scale_y_linear}
\alias{scale_y_linear}
\title{Linear Scale}
\usage{
scale_y_linear(vl, domain, range)
}
\description{
Linear Scale
}
\references{
\href{http://vega.github.io/vega-lite/docs/scale.html}{Vega-Lite Scales spec}
}

