% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.r
\name{scale_y_log}
\alias{scale_y_log}
\title{Log Scale}
\usage{
scale_y_log(vl, domain, range, clamp, nice, zero)
}
\description{
Log Scale
}
\references{
\href{http://vega.github.io/vega-lite/docs/scale.html}{Vega-Lite Scales spec}
}

