% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.r
\name{scale_y_pow}
\alias{scale_y_pow}
\title{Power Scale}
\usage{
scale_y_pow(vl, domain, range, clamp, exp, nice, zero)
}
\description{
Power Scale
}
\references{
\href{http://vega.github.io/vega-lite/docs/scale.html}{Vega-Lite Scales spec}
}

