% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.r
\name{scale_y_threshold}
\alias{scale_y_threshold}
\title{Threshold Scale}
\usage{
scale_y_threshold(vl, domain, range, nice, zero)
}
\description{
Threshold Scale
}
\references{
\href{http://vega.github.io/vega-lite/docs/scale.html}{Vega-Lite Scales spec}
}

