% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.r
\name{timeunit_x}
\alias{timeunit_x}
\title{How to encode x-axis time values}
\usage{
timeunit_x(vl, unit)
}
\arguments{
\item{vl}{Vega-Lite object}

\item{unit}{the property of a channel definition sets the level of specificity
for a temporal field. Currently supported values are 'year', 'yearmonth',
'yearmonthday', 'yearmonthdate', 'yearday', 'yeardate', 'yearmonthdayhours'
and 'yearmonthdayhoursminutes' for non-periodic time units & 'month',
'day', 'date', 'hours', 'minutes', 'seconds', 'milliseconds', 'hoursminutes',
'hoursminutesseconds', 'minutesseconds' and 'secondsmilliseconds' for
periodic time units.}
}
\description{
How to encode x-axis time values
}
\references{
\href{http://vega.github.io/vega-lite/docs/timeunit.html}{Vega-Lite Time Unit}
}

