% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromspec.r
\name{to_spec}
\alias{to_spec}
\title{Convert a spec created with widget idioms to JSON}
\usage{
to_spec(vl)
}
\arguments{
\item{vl}{a Vega-Lite object}
}
\value{
JSON spec
}
\description{
Takes an htmlwidget object and turns it into a JSON Vega-Lite spec
}
\note{
This won't work 100\% of the time yet. \code{calculate}d transforms
      need to be in an array and that requires going back and tagging
      all other single-element vectors with \code{unbox} which is just
      tedious and best saved for a rainy day.
}
\examples{
dat <- jsonlite::fromJSON('[
    {"a": "A","b": 28}, {"a": "B","b": 55}, {"a": "C","b": 43},
    {"a": "D","b": 91}, {"a": "E","b": 81}, {"a": "F","b": 53},
    {"a": "G","b": 19}, {"a": "H","b": 87}, {"a": "I","b": 52}
  ]')

vegalite() \%>\%
  add_data(dat) \%>\%
  encode_x("a", "ordinal") \%>\%
  encode_y("b", "quantitative") \%>\%
  mark_bar() -> chart

to_spec(chart)
}

