% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegalite-package.R
\name{vegalite-exports}
\alias{JS}
\alias{\%>\%}
\alias{saveWidget}
\alias{vegalite-exports}
\title{vegalite exported operators}
\description{
The following functions are imported and then re-exported
from the vegalite package to enable use of the magrittr
pipe operator with no additional library calls.
}

