% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegalite.r
\name{vegalite}
\alias{vegalite}
\title{Create and (optionally) visualize a Vega-Lite spec}
\usage{
vegalite(description = "", renderer = c("svg", "canvas"), export = FALSE,
  source = FALSE, editor = FALSE, viewport_width = NULL,
  viewport_height = NULL, background = NULL, time_format = NULL,
  number_format = NULL)
}
\arguments{
\item{description}{a single element character vector that provides a description of
the plot/spec.}

\item{renderer}{the renderer to use for the view. One of \code{canvas} or
\code{svg} (the default)}

\item{export}{if \code{TRUE} the \emph{"Export as..."} link will
be displayed with the chart.(Default: \code{FALSE}.)}

\item{source}{if \code{TRUE} the \emph{"View Source"} link will be displayed
with the chart. (Default: \code{FALSE}.)}

\item{editor}{if \code{TRUE} the \emph{"Open in editor"} link will be
displayed with the cahrt. (Default: \code{FALSE}.)}

\item{viewport_width, viewport_height}{height and width of the overall
visualziation viewport. This is the overall area reserved for the
plot. You can leave these \code{NULL} and use \code{\link{cell_size}}
instead but you will want to configure both when making faceted plots.}

\item{background}{plot background color. If \code{NULL} the background will be transparent.}

\item{time_format}{the default time format pattern for text and labels of
axes and legends (in the form of \href{https://github.com/mbostock/d3/wiki/Time-Formatting}{D3 time format pattern}).
Default: \code{\%Y-\%m-\%d}}

\item{number_format}{the default number format pattern for text and labels of
axes and legends (in the form of
\href{https://github.com/mbostock/d3/wiki/Formatting}{D3 number format pattern}).
Default: \code{s}}
}
\description{
Create and (optionally) visualize a Vega-Lite spec
}
\examples{
dat <- jsonlite::fromJSON('[
    {"a": "A","b": 28}, {"a": "B","b": 55}, {"a": "C","b": 43},
    {"a": "D","b": 91}, {"a": "E","b": 81}, {"a": "F","b": 53},
    {"a": "G","b": 19}, {"a": "H","b": 87}, {"a": "I","b": 52}
  ]')

vegalite() \%>\%
  add_data(dat) \%>\%
  encode_x("a", "ordinal") \%>\%
  encode_y("b", "quantitative") \%>\%
  mark_bar()
}
\references{
\href{http://vega.github.io/vega-lite/docs/config.html#top-level-config}{Vega-Lite top-level config}
}

