% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.r
\name{vegaliteOutput}
\alias{vegaliteOutput}
\title{Widget output function for use in Shiny}
\usage{
vegaliteOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

