% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.r
\encoding{UTF-8}
\name{view_config}
\alias{view_config}
\alias{view_size}
\title{view aesthetics}
\usage{
view_config(vl, width = 200, height = 200, clip = FALSE, fill = NULL,
  fill_opacity = NULL, stroke = NULL, stroke_opacity = NULL,
  stroke_width = NULL, stroke_dash = NULL, stroke_dash_offset = NULL)

view_size(vl, width = 200, height = 200)
}
\arguments{
\item{vl}{Vega-Lite object}

\item{width, height}{width and height property of the view configuration determine
the width of a visualization with a continuous x-scale and the height of
a visualization with a continuous y-scale respectively. Visit the
URL in the References section for more information.}

\item{clip}{\code{FALSE} by default, whether the view should be clipped.}

\item{fill}{fill color}

\item{fill_opacity}{\code{0.0}-\code{1.0}}

\item{stroke}{stroke color}

\item{stroke_opacity}{\code{0.0}-\code{1.0}}

\item{stroke_width}{stroke of the width in pixels}

\item{stroke_dash}{an array of alternating stroke, space lengths for creating
dashed or dotted lines.}

\item{stroke_dash_offset}{the offset (in pixels) into which to begin drawing with the stroke dash array.}
}
\description{
At its core, a Vega-Lite specification describes a single plot. This function
works to format those views with all of the options avaiable under
\href{https://vega.github.io/vega-lite/docs/config.html#view-config}{config.view.*}
When a facet channel is added, the visualization is faceted into a trellis
plot, which contains multiple plots. Each plot in either a single plot or a
trellis plot is called a view. View configuration allows us to customize each
individual single plot and each plot in a trellis plot.
}
\examples{
vegalite() \%>\%
  view_size(300, 200) \%>\%
  add_data("https://vega.github.io/vega-editor/app/data/unemployment-across-industries.json") \%>\%
  encode_x("date", "temporal") \%>\%
  encode_y("count", "quantitative", aggregate="sum") \%>\%
  encode_color("series", "nominal") \%>\%
  scale_color_nominal_vl(scheme="category20b") \%>\%
  timeunit_x("yearmonth") \%>\%
  scale_x_time_vl(nice="month") \%>\%
  axis_x(format="\%Y", labelAngle=0) \%>\%
  mark_area()
}
\references{
\href{http://vega.github.io/vega-lite/docs/config.html#view-config}{Vega-Lite view spec}
}
