% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phantom.r
\name{widgetThumbnail}
\alias{widgetThumbnail}
\title{Make a thumbnail for an htmlwidget panel}
\usage{
widgetThumbnail(p, thumbPath)
}
\arguments{
\item{p}{htmlwidget object}

\item{thumbPath}{where to save thumbnail file}
}
\description{
Make a thumbnail for an htmlwidget panel
}

