% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.r
\name{axis_facet_row}
\alias{axis_facet_row}
\title{Facet axis setttings}
\usage{
axis_facet_row(vl, axisWidth = 0, layer = NULL, offset = NULL,
  grid = FALSE, labels = TRUE, labelAngle = NULL, labelAlign = NULL,
  labelBaseline = NULL, labelMaxLength = 25, shortTimeLabels = NULL,
  subdivide = NULL, ticks = NULL, tickPadding = NULL, tickSize = 0,
  tickSizeMajor = NULL, tickSizeMinor = NULL, tickSizeEnd = NULL,
  title = "", titleOffset = NULL, titleMaxLength = NULL,
  characterWidth = 6, orient = NULL, format = NULL, remove = FALSE)
}
\description{
Facet axis setttings
}
\references{
\href{http://vega.github.io/vega-lite/docs/axis.html}{Vega-List Axis spec}
}

