% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.r
\encoding{UTF-8}
\name{axis_facet_row}
\alias{axis_facet_row}
\title{General axis setttings (row facets)}
\usage{
axis_facet_row(vl, axisWidth = 0, layer = NULL, offset = NULL,
  grid = FALSE, labels = TRUE, labelAngle = NULL, labelAlign = NULL,
  labelBaseline = NULL, labelMaxLength = 25, shortTimeLabels = NULL,
  subdivide = NULL, ticks = NULL, tickPadding = NULL, tickSize = 0,
  tickSizeMajor = NULL, tickSizeMinor = NULL, tickSizeEnd = NULL,
  title = "", titleOffset = NULL, titleMaxLength = NULL,
  characterWidth = 6, orient = NULL, format = NULL, remove = FALSE)
}
\arguments{
\item{vl}{Vega-Lite object}

\item{axisWidth, layer, offset, grid, labels, labelAngle, labelAlign, labelBaseline}{see \href{http://vega.github.io/vega-lite/docs/axis.html}{axis docs} &
\href{https://github.com/vega/vega-lite/blob/master/src/axis.ts}{axis base config}}

\item{labelMaxLength, shortTimeLabels, subdivide, ticks, tickPadding, tickSize}{see \href{http://vega.github.io/vega-lite/docs/axis.html}{axis docs} &
\href{https://github.com/vega/vega-lite/blob/master/src/axis.ts}{axis base config}}

\item{tickSizeMajor, tickSizeMinor, tickSizeEnd, title, titleOffset, titleMaxLength}{see \href{http://vega.github.io/vega-lite/docs/axis.html}{axis docs} &
\href{https://github.com/vega/vega-lite/blob/master/src/axis.ts}{axis base config}}

\item{characterWidth, orient, format, remove}{see \href{http://vega.github.io/vega-lite/docs/axis.html}{axis docs} &
\href{https://github.com/vega/vega-lite/blob/master/src/axis.ts}{axis base config}}
}
\description{
Axes provide axis lines, ticks and labels to convey how a spatial range represents
a data range. Simply put, axes visualize scales. \cr
\cr
By default, Vega-Lite automatically creates axes for x, y, row, and column channels
when they are encoded. Axis can be customized via the axis property of a channel
definition.
}
\references{
\href{http://vega.github.io/vega-lite/docs/axis.html}{Vega-List Axis spec}
}

