% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.r
\name{axis_x}
\alias{axis_x}
\title{General axis setttings}
\usage{
axis_x(vl, axisWidth = NULL, layer = NULL, offset = NULL, grid = NULL,
  labels = TRUE, labelAngle = NULL, labelAlign = NULL,
  labelBaseline = NULL, labelMaxLength = 25, shortTimeLabels = NULL,
  subdivide = NULL, ticks = NULL, tickPadding = NULL, tickSize = NULL,
  tickSizeMajor = NULL, tickSizeMinor = NULL, tickSizeEnd = NULL,
  title = "", titleOffset = NULL, titleMaxLength = NULL,
  characterWidth = 6, orient = NULL, format = NULL, remove = FALSE)
}
\description{
General axis setttings
}
\references{
\href{http://vega.github.io/vega-lite/docs/axis.html}{Vega-List Axis spec}
}

