% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.r
\name{axis_x}
\alias{axis_x}
\title{General axis setttings}
\usage{
axis_x(vl, layer, offset, orient, axisWidth, format, labelAngle)
}
\description{
General axis setttings
}
\references{
\href{http://vega.github.io/vega-lite/docs/axis.html}{Vega-List Axis spec}
}

