% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.r
\name{calculate}
\alias{calculate}
\title{Derive new fields}
\usage{
calculate(vl, field, expr)
}
\arguments{
\item{vl}{Vega-Lite object created by \code{\link{vegalite}}}

\item{field}{the field name in which to store the computed value.}

\item{expr}{a tring containing an expression for the formula. Use the variable
\code{"datum"} to refer to the current data object.}
}
\description{
Derive new fields
}

