% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.r
\name{encode_size}
\alias{encode_size}
\title{Encode size "channel"}
\usage{
encode_size(vl, field = NULL, type, value = NULL)
}
\arguments{
\item{vl}{Vega-Lite object created by \code{\link{vegalite}}}

\item{field}{single element character vector naming the column}

\item{type}{the encoded field’s type of measurement. This can be either a full type
name (\code{quantitative}, \code{temporal}, \code{ordinal}, and \code{nominal})
or an initial character of the type name (\code{Q}, \code{T}, \code{O}, \code{N}).
This property is case insensitive. If \code{auto} is used, the type will
be guessed (so you may want to actually specify it if you want consistency).}

\item{value}{if a field is not specified, constant values for the properties
(e.g., color, size) can be also set directly with the channel definition’s value property.}
}
\description{
Encode size "channel"
}
\examples{
vegalite() \%>\%
  add_data("https://vega.github.io/vega-editor/app/data/cars.json") \%>\%
  encode_x("Horsepower", "quantitative") \%>\%
  encode_y("Miles_per_Gallon", "quantitative") \%>\%
  encode_size("Acceleration", "quantitative") \%>\%
  mark_point()
}
\references{
\href{http://vega.github.io/vega-lite/docs/encoding.html}{Vega-Lite Encoding spec}
}

