% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.r
\encoding{UTF-8}
\name{legend_color}
\alias{legend_color}
\title{Legend settings (color)}
\usage{
legend_color(vl, ...)
}
\arguments{
\item{vl}{a Vega-Lite object}

\item{orient, offset, values}{see \href{https://vega.github.io/vega-lite/docs/legend.html#general}{legend docs}}

\item{title}{the title for the legend.}

\item{format, labelAlign, labelBaseline, labelColor, labelFont, labelFontSize, short_time_labels}{see \href{https://vega.github.io/vega-lite/docs/legend.html#labels}{legend docs}}

\item{symbolColor, symbolShape, symbolSize, symbolStrokeWidth}{see \href{https://vega.github.io/vega-lite/docs/legend.html#symbols}{legend docs}}

\item{title, titleColor, titleFont, titleFontSize, titleFontWeight}{see \href{https://vega.github.io/vega-lite/docs/legend.html#title}{legend docs}}

\item{remove}{if \code{TRUE}, there will be no legend for this aesthetic.}
}
\description{
Legend settings (color)
}

