% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.r
\name{mark_area}
\alias{mark_area}
\title{Area mark}
\usage{
mark_area(vl, orient = NULL, stack = NULL, interpolate = NULL,
  tension = NULL)
}
\arguments{
\item{vl}{def}

\item{orient}{def}

\item{stack}{def}

\item{interpolate}{def}

\item{tension}{def}
}
\description{
Area mark
}
\references{
\href{http://vega.github.io/vega-lite/docs/mark.html}{Vega-Lite Mark spec}
}

