% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.r
\encoding{UTF-8}
\name{mark_bar}
\alias{mark_bar}
\title{Bar mark}
\usage{
mark_bar(vl, orient = NULL, stack = NULL, size = NULL, opacity = NULL)
}
\arguments{
\item{vl}{Vega-Lite object}

\item{orient}{the orientation of the axis. One of top or bottom for y and row
channels, and left or right for x and column channels. Default value:
x axis is placed on the bottom, y axis is placed on the left, column”s
x-axis is placed on the top, row”s y-axis is placed on the right.}

\item{stack}{stacking modes for bar and area marks. \code{zero} - stacking
with baseline offset at zero value of the scale (for creating typical stacked
bar and area chart). \code{normalize} - stacking with normalized domain (for
creating normalized stacked bar and area chart). \code{center} - stacking
with center baseline (for streamgraph). \code{none} - No-stacking. This will
produces layered bar and area chart.}

\item{size}{The pixel area each the point. For example: in the case of circles,
the radius is determined in part by the square root of the size value.}

\item{opacity}{\code{0.0}-\code{1.0}}
}
\description{
A bar mark represents each data point as a rectangle, where the length is
mapped to a quantitative scale.
}
\examples{
dat <- jsonlite::fromJSON('[
    {"a": "A","b": 28}, {"a": "B","b": 55}, {"a": "C","b": 43},
    {"a": "D","b": 91}, {"a": "E","b": 81}, {"a": "F","b": 53},
    {"a": "G","b": 19}, {"a": "H","b": 87}, {"a": "I","b": 52}
  ]')

vegalite() \%>\%
  add_data(dat) \%>\%
  encode_x("a", "ordinal") \%>\%
  encode_y("b", "quantitative") \%>\%
  mark_bar()
}
\references{
\href{http://vega.github.io/vega-lite/docs/mark.html}{Vega-Lite Mark spec}
}

