% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.r
\name{mark_circle}
\alias{mark_circle}
\title{Circle mark}
\usage{
mark_circle(vl, size = 30, opacity = NULL)
}
\arguments{
\item{size}{The pixel area each the point. For example: in the case of circles,
the radius is determined in part by the square root of the size value.
Default value: 30}

\item{opacity}{\code{0.0}-\code{1.0}}
}
\description{
circle and square marks are similar to point mark, except that (1) the shape
value is always set to circle or square (2) they are filled by default.
}
\examples{
vegalite() \%>\%
  add_data("https://vega.github.io/vega-editor/app/data/cars.json") \%>\%
  encode_x("Horsepower", "quantitative") \%>\%
  encode_y("Miles_per_Gallon", "quantitative") \%>\%
  mark_circle()
}
\references{
\href{http://vega.github.io/vega-lite/docs/mark.html}{Vega-Lite Mark spec}
}

