% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.r
\name{mark_line}
\alias{mark_line}
\title{Line mark}
\usage{
mark_line(vl, orient = NULL, interpolate = NULL, tension = NULL,
  opacity = NULL)
}
\arguments{
\item{vl}{def}

\item{orient}{def}

\item{interpolate}{def}

\item{tension}{def}

\item{opacity}{0.0-1.0}
}
\description{
Line mark
}
\references{
\href{http://vega.github.io/vega-lite/docs/mark.html}{Vega-Lite Mark spec}
}

