% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.r
\name{mark_point}
\alias{mark_point}
\title{Point mark}
\usage{
mark_point(vl)
}
\description{
Point mark
}
\examples{
vegalite() \%>\%
  add_data("https://vega.github.io/vega-editor/app/data/cars.json") \%>\%
  encode_x("Horsepower", "quantitative") \%>\%
  encode_y("Miles_per_Gallon", "quantitative") \%>\%
  mark_point()
}
\references{
\href{http://vega.github.io/vega-lite/docs/mark.html}{Vega-Lite Mark spec}
}

