% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.r
\name{mark_square}
\alias{mark_square}
\title{Square mark}
\usage{
mark_square(vl, size = NULL, opacity = NULL)
}
\arguments{
\item{size}{The pixel area each the point. For example: in the case of circles,
the radius is determined in part by the square root of the size value.}

\item{opacity}{\code{0.0}-\code{1.0}}
}
\description{
Circle and square marks are similar to point mark, except that (1) the shape
value is always set to circle or square (2) they are filled by default.
}
\references{
\href{http://vega.github.io/vega-lite/docs/mark.html}{Vega-Lite Mark spec}
}

