% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.r
\name{mark_text}
\alias{mark_text}
\title{Text mark}
\usage{
mark_text(vl, opacity = NULL)
}
\arguments{
\item{opacity}{0.0-1.0}
}
\description{
A text mark represents each data point with a text instead of a point.
}
\references{
\href{http://vega.github.io/vega-lite/docs/mark.html}{Vega-Lite Mark spec}
}

