% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.r
\name{mark_tick}
\alias{mark_tick}
\title{Tick mark}
\usage{
mark_tick(vl, orient = NULL, size = NULL, thickness = 1, opacity = NULL)
}
\arguments{
\item{vl}{def}

\item{orient}{def}

\item{size}{def}

\item{thickness}{def}

\item{opacity}{0.0-1.0}
}
\description{
Tick mark
}
\examples{
vegalite() \%>\%
  add_data("https://vega.github.io/vega-editor/app/data/cars.json") \%>\%
  encode_x("Horsepower", "quantitative") \%>\%
  encode_y("Cylinders", "ordinal") \%>\%
  mark_tick()
}
\references{
\href{http://vega.github.io/vega-lite/docs/mark.html}{Vega-Lite Mark spec}
}

