% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.r
\name{renderVegalite}
\alias{renderVegalite}
\alias{renderVegaliteSpec}
\title{Widget render function for use in Shiny}
\usage{
renderVegalite(expr, env = parent.frame(), quoted = FALSE)

renderVegaliteSpec(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{expr to render}

\item{env}{evaluation environemnt}

\item{quoted}{quote expression?}
}
\description{
Widget render function for use in Shiny
}
