% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.r
\encoding{UTF-8}
\name{scale_shape}
\alias{scale_shape}
\title{Shape Scale}
\usage{
scale_shape(vl, domain, range)
}
\arguments{
\item{vl}{Vega-Lite object}

\item{domain}{Custom domain values. For quantitative data, this can take the
form of a two-element array with minimum and maximum values.}

\item{range}{The range of the scale represents the set of output visual values.
Vega-Lite automatically determines appropriate range based on the scale’s
channel and type, but range property can be provided to customize range
values.}
}
\description{
Shape Scale
}
\references{
\href{http://vega.github.io/vega-lite/docs/scale.html}{Vega-Lite Scales spec}
}

