% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.r
\name{scale_x_ordinal}
\alias{scale_x_ordinal}
\title{Ordinal Scale}
\usage{
scale_x_ordinal(vl, domain, range)
}
\description{
Ordinal Scale
}
\references{
\href{http://vega.github.io/vega-lite/docs/scale.html}{Vega-Lite Scales spec}
}

