% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.r
\encoding{UTF-8}
\name{scale_x_ordinal}
\alias{scale_x_ordinal}
\title{Ordinal Scale}
\usage{
scale_x_ordinal(vl, band_size = NULL, padding = NULL)
}
\arguments{
\item{vl}{Vega-Lite object}

\item{band_size}{band size}

\item{padding}{padding}
}
\description{
Ordinal Scale
}
\references{
\href{http://vega.github.io/vega-lite/docs/scale.html}{Vega-Lite Scales spec}
}

