% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.r
\name{scale_x_pow}
\alias{scale_x_pow}
\title{Quantitative Scale}
\usage{
scale_x_pow(vl, domain, range)
}
\description{
Quantitative Scale
}
\references{
\href{http://vega.github.io/vega-lite/docs/scale.html}{Vega-Lite Scales spec}
}

