% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.r
\name{validate_vl}
\alias{validate_vl}
\title{validate_vl}
\usage{
validate_vl(vl, message = TRUE)
}
\arguments{
\item{vl}{Vega-Lite object}

\item{message}{if invalid, print message?}
}
\value{
\code{TRUE} if valid, \code{FALSE} if not
}
\description{
Validate the Vega-Lite widget against the schema
}
\examples{
dat <- jsonlite::fromJSON('[
    {"a": "A","b": 28}, {"a": "B","b": 55}, {"a": "C","b": 43},
    {"a": "D","b": 91}, {"a": "E","b": 81}, {"a": "F","b": 53},
    {"a": "G","b": 19}, {"a": "H","b": 87}, {"a": "I","b": 52}
  ]')

vl <- vegalite() \%>\%
  add_data(dat) \%>\%
  encode_x("a", "ordinal") \%>\%
  encode_y("b", "quantitative") \%>\%
  mark_bar()

validate_vl(vl)
}
